+++ src/main/scala/dagger/ml/Classifier.scala
@@ -8,28 +8,16 @@ import gnu.trove.map.hash.THashMap
  */
 abstract class MultiClassClassifier[T] {
 
-  def dotMap(v1: collection.Map[Int, Double], v2: collection.Map[Int, Double]): Double = {
-    v1.foldLeft(0.0){ case(sum, (f,v)) =>
-      sum + v * v2.getOrElse(f, 0.0)
+  def dotMap(v1: collection.Map[Int, Float], v2: collection.Map[Int, Float]): Float = {
+    v1.foldLeft(0.0f) {
+      case (sum, (f, v)) =>
+        sum + v * v2.getOrElse(f, 0.0f)
     }
   }
 
   def predict(instance: Instance[T]): Prediction[T]
 
-  def weightOf(a: T, p: Int): Double
+  def weightOf(a: T, p: Int): Float
 
   def writeToFile(filename: String): Unit
 }
-
-
-
-
-
-
-
-//  def dotMap(v1: THashMap[Int, Double], v2: THashMap[Int, Double]): Double = {
-//    var sum = 0.0
-//    v1.foldLeft(0.0){ case(sum, (f,v)) =>
-//      sum + v * v2.getOrElse(f, 0.0)
-//    }
-//  }
